style: Navigation content under trigger
